Refactor app into controller hook and board components
